- Add setup.py - Seperate examples - Make it possible to choose pyside or pyqt4 by using different imports
