+++ pygletwidget/pyside/__init__.py
@@ -1,0 +1,5 @@
+import pygletwidget
+pygletwidget.QT_FLAVOUR="PYSIDE"
+
+from pygletwidget._qpygletwidget import *
+
