+++ pygletwidget/qt4/__init__.py
@@ -1,0 +1,4 @@
+import pygletwidget
+pygletwidget.QT_FLAVOUR="QT4"
+
+from pygletwidget._qpygletwidget import *
